+++ packages/sdk/src/openapibuilder/index.test.ts
@@ -167,18 +167,16 @@ describe('OpenAPI builder', () => {
 		expect(querySpec.get?.parameters?.[0].schema).toStrictEqual({ type: 'string' });
 		const queryResponse = querySpec.get?.responses['200'];
 		expect(queryResponse).toBeDefined();
-		expect(queryResponse?.content['application/json'].schema).toStrictEqual(operations[0].ResponseSchema);
+		expect(queryResponse?.content['application/json'].schema).toEqual(operations[0].ResponseSchema);
 
 		const mutationSpec = result.paths['/MutationPath'];
 		expect(mutationSpec).toBeDefined();
 		expect(mutationSpec.get).toBeUndefined();
 		expect(mutationSpec.post?.operationId).toBe('Mutation');
-		expect(mutationSpec.post?.requestBody?.content['application/json'].schema).toStrictEqual(
-			operations[1].VariablesSchema
-		);
+		expect(mutationSpec.post?.requestBody?.content['application/json'].schema).toEqual(operations[1].VariablesSchema);
 		const mutationResponse = mutationSpec.post?.responses['200'];
 		expect(mutationResponse).toBeDefined();
-		expect(mutationResponse?.content['application/json'].schema).toStrictEqual(operations[1].ResponseSchema);
+		expect(mutationResponse?.content['application/json'].schema).toEqual(operations[1].ResponseSchema);
 	});
 
 	test('skip internal operations', async () => {
@@ -300,7 +298,163 @@ describe('OpenAPI builder', () => {
 		expect((inputSchema?.properties?.['data'] as JSONSchema)?.$ref).toBe('#/components/schemas/Person');
 		const personDefinition = result?.components?.schemas?.['Person'];
 		expect(personDefinition).toBeDefined();
-		expect(personDefinition).toStrictEqual(personSchema?.definitions?.['Person']);
+		expect(personDefinition).toEqual(personSchema?.definitions?.['Person']);
+	});
+
+	test('rename different types with same name', () => {
+		const userSchema: JSONSchema = {
+			type: 'object',
+			properties: {
+				name: { type: 'string' },
+				surname: { type: 'string' },
+			},
+			required: ['name', 'surname'],
+			additionalProperties: false,
+		};
+		const user2Schema: JSONSchema = {
+			type: 'object',
+			properties: {
+				username: { type: 'string' },
+				age: { type: 'number' },
+			},
+			required: ['username', 'age'],
+			additionalProperties: false,
+		};
+		const user3Schema: JSONSchema = {
+			type: 'object',
+			properties: {
+				username: { type: 'string' },
+				city: { type: 'string' },
+			},
+			required: ['username', 'city'],
+			additionalProperties: false,
+		};
+
+		const operations = [
+			{
+				Name: 'GetUserA',
+				PathName: 'users/getA',
+				OperationType: OperationType.QUERY,
+				ExecutionEngine: OperationExecutionEngine.ENGINE_GRAPHQL,
+				VariablesSchema: emptySchema,
+				ResponseSchema: {
+					type: 'object',
+					properties: {
+						data: {
+							$ref: '#/definitions/User',
+						},
+					},
+					additionalProperties: false,
+					definitions: {
+						User: userSchema,
+					},
+				},
+			},
+			{
+				Name: 'GetUserB',
+				PathName: 'users/getB',
+				OperationType: OperationType.QUERY,
+				ExecutionEngine: OperationExecutionEngine.ENGINE_GRAPHQL,
+				VariablesSchema: emptySchema,
+				ResponseSchema: {
+					type: 'object',
+					properties: {
+						data: {
+							$ref: '#/definitions/User',
+						},
+					},
+					additionalProperties: false,
+					definitions: {
+						User: user2Schema,
+					},
+				},
+			},
+			{
+				Name: 'GetUserC',
+				PathName: 'users/getC',
+				OperationType: OperationType.QUERY,
+				ExecutionEngine: OperationExecutionEngine.ENGINE_GRAPHQL,
+				VariablesSchema: emptySchema,
+				ResponseSchema: {
+					type: 'object',
+					properties: {
+						data: {
+							$ref: '#/definitions/User',
+						},
+					},
+					additionalProperties: false,
+					definitions: {
+						User: user3Schema,
+					},
+				},
+			},
+			{
+				Name: 'GetUserD',
+				PathName: 'users/getD',
+				OperationType: OperationType.QUERY,
+				ExecutionEngine: OperationExecutionEngine.ENGINE_GRAPHQL,
+				VariablesSchema: emptySchema,
+				ResponseSchema: {
+					type: 'object',
+					properties: {
+						data: {
+							$ref: '#/definitions/User',
+						},
+					},
+					additionalProperties: false,
+					definitions: {
+						User: user2Schema,
+					},
+				},
+			},
+			{
+				Name: 'SetUserE',
+				PathName: 'users/setE',
+				OperationType: OperationType.MUTATION,
+				ExecutionEngine: OperationExecutionEngine.ENGINE_GRAPHQL,
+				VariablesSchema: {
+					type: 'object',
+					properties: {
+						data: {
+							$ref: '#/definitions/User',
+						},
+					},
+					additionalProperties: false,
+					definitions: {
+						User: user3Schema,
+					},
+				},
+				ResponseSchema: emptySchema,
+			},
+		] as unknown as GraphQLOperation[];
+
+		const result = build(operations);
+		const operationA = result.paths['/users/getA'].get;
+		const operationB = result.paths['/users/getB'].get;
+		const operationC = result.paths['/users/getC'].get;
+		const operationD = result.paths['/users/getD'].get;
+		const operationE = result.paths['/users/setE'].post;
+
+		const responseASchema = operationA?.responses?.['200']?.content?.['application/json']?.schema;
+		expect((responseASchema?.properties?.['data'] as JSONSchema7)?.$ref).toBe('#/components/schemas/User');
+
+		const responseBSchema = operationB?.responses?.['200']?.content?.['application/json']?.schema;
+		expect((responseBSchema?.properties?.['data'] as JSONSchema7)?.$ref).toBe('#/components/schemas/User_2');
+
+		const responseCSchema = operationC?.responses?.['200']?.content?.['application/json']?.schema;
+		expect((responseCSchema?.properties?.['data'] as JSONSchema7)?.$ref).toBe('#/components/schemas/User_3');
+
+		const responseDSchema = operationD?.responses?.['200']?.content?.['application/json']?.schema;
+		expect((responseDSchema?.properties?.['data'] as JSONSchema7)?.$ref).toBe('#/components/schemas/User_2');
+
+		const inputESchema = operationE?.requestBody?.content?.['application/json']?.schema;
+		expect((inputESchema?.properties?.['data'] as JSONSchema7)?.$ref).toBe('#/components/schemas/User_3');
+
+		const schemas = result?.components?.schemas;
+
+		expect(schemas?.['User']).toEqual(userSchema);
+		expect(schemas?.['User_2']).toEqual(user2Schema);
+		expect(schemas?.['User_3']).toEqual(user3Schema);
 	});
 
 	test('OpenAPI Builder', async () => {