+++ pkg/apihandler/apihandler.go
@@ -20,9 +20,9 @@ import (
 	"github.com/gorilla/mux"
 	"github.com/gorilla/securecookie"
 	"github.com/hashicorp/go-uuid"
-	"github.com/mattbaird/jsonpatch"
 	"github.com/tidwall/gjson"
 	"github.com/tidwall/sjson"
+	"github.com/wI2L/jsondiff"
 	"go.uber.org/zap"
 
 	"github.com/wundergraph/graphql-go-tools/pkg/ast"
@@ -171,7 +171,7 @@ func (r *Builder) BuildAndMountApiHandler(ctx context.Context, router *mux.Route
 		return streamClosers, fmt.Errorf("authentication config missing")
 	}
 
-	planConfig, err := r.loader.Load(*api.EngineConfiguration, api.Options.ServerUrl)
+	planConfig, err := r.loader.Load(api.EngineConfiguration, api.Options.ServerUrl)
 	if err != nil {
 		return streamClosers, err
 	}
@@ -1081,7 +1081,7 @@ func (h *QueryHandler) handleLiveQuery(r *http.Request, w http.ResponseWriter, c
 			if wgParams.UseJsonPatch && lastData.Len() != 0 {
 				last := lastData.Bytes()
 				current := currentData.Bytes()
-				patch, err := jsonpatch.CreatePatch(last, current)
+				patch, err := jsondiff.CompareJSON(last, current)
 				if err != nil {
 					requestLogger.Error("HandleLiveQueryEvent could not create json patch", zap.Error(err))
 					continue
@@ -1448,7 +1448,7 @@ func (f *httpFlushWriter) Flush() {
 
 	if f.useJsonPatch && f.lastMessage.Len() != 0 {
 		last := f.lastMessage.Bytes()
-		patch, err := jsonpatch.CreatePatch(last, resp)
+		patch, err := jsondiff.CompareJSON(last, resp)
 		if err != nil {
 			if f.logger != nil {
 				f.logger.Error("subscription json patch", zap.Error(err))