+++ pkg/telemetry/data_sources.go
@@ -1,12 +1,9 @@
 package telemetry
 
 import (
-	"encoding/json"
 	"fmt"
 	"net"
 	"net/url"
-	"os"
-	"path/filepath"
 
 	"github.com/wundergraph/wundergraph/pkg/loadvariable"
 	"github.com/wundergraph/wundergraph/pkg/wgpb"
@@ -89,15 +86,8 @@ func dataSourceMetric(dataSourceTag string, urlVariable *wgpb.ConfigurationVaria
 }
 
 func DataSourceMetrics(wunderGraphDir string) ([]*Metric, error) {
-	configFile := filepath.Join(wunderGraphDir, "generated", "wundergraph.config.json")
-	f, err := os.Open(configFile)
+	wgConfig, err := readWunderGraphConfig(wunderGraphDir)
 	if err != nil {
-		return nil, err
-	}
-	defer f.Close()
-	dec := json.NewDecoder(f)
-	var wgConfig wgpb.WunderGraphConfiguration
-	if err := dec.Decode(&wgConfig); err != nil {
 		return nil, err
 	}
 	var metrics []*Metric