+++ packages/golang-client/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.12](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.6.11...@wundergraph/golang-client@0.6.12) (2023-07-06)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
+## [0.6.11](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.6.10...@wundergraph/golang-client@0.6.11) (2023-07-05)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
+## [0.6.10](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.6.9...@wundergraph/golang-client@0.6.10) (2023-06-22)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
+## [0.6.9](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.6.8...@wundergraph/golang-client@0.6.9) (2023-06-21)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
 ## [0.6.8](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.6.7...@wundergraph/golang-client@0.6.8) (2023-06-09)
 
 **Note:** Version bump only for package @wundergraph/golang-client